+++ JsonSubTypes.Text.Json.Tests.Shared/SharedAnimal.cs
@@ -1,7 +1,7 @@
 namespace JsonSubTypes.Text.Json.Tests.Shared
 {
     [JsonSubTypeConverter(typeof(JsonSubtypes<SharedAnimal>), "Kind")]
-    [JsonSubTypesTypeResolution("JsonSubTypes.Text.Json.Tests.Plugin")]
+    [KnownSubTypeOtherAssembly("JsonSubTypes.Text.Json.Tests.Plugin")]
     public class SharedAnimal
     {
         public string? Kind { get; set; }